Moved internal structs
